[Refactor] Refactor name / Call Type
