+++ app/src/main/java/org/sopt/sample/application/ReqresApi.kt
@@ -1,10 +1,10 @@
 package org.sopt.sample.application
 
-import org.sopt.sample.data.model.dto.ResponseReqresList
+import org.sopt.sample.data.model.dto.ResponseReqresListDTO
 import retrofit2.Call
-import retrofit2.http.POST
+import retrofit2.http.GET
 
 interface ReqresApi {
-    @POST("api/users?page=2")
-    fun signup(): Call<ResponseReqresList>
+    @GET("api/users?page=2")
+    fun getList(): Call<ResponseReqresListDTO>
 }
